Revert "modernize the application"
