+++ package.json
@@ -10,32 +10,32 @@
     "generate": "drizzle-kit generate:pg"
   },
   "dependencies": {
-    "@auth/drizzle-adapter": "^1.4.1",
+    "@auth/drizzle-adapter": "^0.8.0",
     "@react-spring/web": "^9.7.3",
-    "@tanstack/react-query": "^5.51.1",
+    "@tanstack/react-query": "^5.28.4",
     "@trpc/client": "11.0.0-next-beta.318",
     "@trpc/server": "11.0.0-next-beta.318",
-    "@use-gesture/react": "^10.3.1",
-    "drizzle-orm": "^0.32.0",
-    "next": "15.0.0-rc.0",
+    "@use-gesture/react": "^10.3.0",
+    "drizzle-orm": "^0.30.3",
+    "next": "14.1.3",
     "next-auth": "^4.24.7",
-    "postgres": "^3.4.4",
-    "react": "19.0.0-rc-8b08e99e-20240713",
-    "react-dom": "19.0.0-rc-8b08e99e-20240713",
+    "postgres": "^3.4.3",
+    "react": "^18",
+    "react-dom": "^18",
     "superjson": "^2.2.1",
-    "zod": "^3.23.8"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.13",
-    "@types/node": "^20.14.10",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "autoprefixer": "^10.4.19",
-    "drizzle-kit": "^0.23.0",
-    "eslint": "^9.7.0",
-    "eslint-config-next": "14.2.5",
-    "postcss": "^8.4.39",
-    "tailwindcss": "^3.4.5",
-    "typescript": "^5.5.3"
+    "@tailwindcss/typography": "^0.5.10",
+    "@types/node": "^20.11.29",
+    "@types/react": "^18.2.67",
+    "@types/react-dom": "^18.2.22",
+    "autoprefixer": "^10.4.18",
+    "drizzle-kit": "^0.20.14",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "14.1.3",
+    "postcss": "^8.4.36",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.4.2"
   }
 }
