--- src/app/game/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return <h1>Loading...</h1>;
-}
