+++ src/app/globals.css
@@ -1,7 +1,3 @@
-@view-transition {
-  navigation: auto;
-}
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
